Add configuration file parser.
The file format is a simple, plain-text, key=value format, with support for
nested sections.  Keys and values are strings, but integer- (with proper
error checking and multiple numerical base support) and boolean- (with
various synonyms for boolean values) are provided.

The usage is pretty simple, and will be better illustrated by the next
commit, which wires the configuration parsing to lwan.