Wire the configuration file reader.
Various changes were needed to support this.

First, handlers are now supposed to implement a `init_from_hash()` function,
which takes a hash table of key=value pairs and initializes the module.  See
the file serving handler for an example.

Tries now have a node-freeing function.

Lwan now links with -ldl so that it can `dlsym()` itself to find the
required symbols.  This might (and probably will) change in the future.

The configuration file must have the same name as the executable, but with a
`.conf` suffix.  It is nowadays being opened from the current directory, but
this might change in the future (maybe respecting XDG config spec or
something of similar nature).