Can't change config settings via struct initializer
The freegeoip example tries to set some config settings directly via a
struct literal, but this does not appear to be a supported way to do
that since the first thing lwan_init() does is overwrite it with its own
defaults. Someone looking through the examples might be misled by this.