Use alternative method calling syntax to obtain query parameters
Initially, accessing query parameters would create a userdata with a
metatable implementing __index to fetch the variable. This proved to be
too much too complex, so a simple function with a string argument is
what's going to be. Just use an alternative syntax to sort of make it
look like an associative array subscription.