Perform proper template lexing+parsing
This implements the same kind of state-machine-with-function-pointers
as shown by Rob Pike on this video[1]. The parser implemented here
should be 100% compatible with the one it substitutes, but it is a lot
easier to understand and extend.

It should now be easier to change the chunk list to be a chunk array.

Closes #84.

[1] https://www.youtube.com/watch?v=HxaD_trXwRE