Revert commits aa0415c and e1d2ba99
It's dangerous to call coro_run_deferred() unless the coroutine is
being destroyed and/or reset.  There might be resources still being
used by the coroutine at that point, which would lead to serious
consequences.

The idea of "running the garbage collector" is sound, however, but
there must be some flag of sorts to mark that a resource isn't used.
Kind of a generational GC (each generation being a coroutine reset).