Work around cross-compilation issues with mimegen
This moves the mimegen utility to be an external project, and so it's
built and executed on the host.  I'm not sure if this is the best
approach, specially since it's essentially calls CMake and make again,
and on some platforms that might break.  But since this is working for
me, both on Linux and FreeBSD, I'm pushing this anyway.

Also, check if `-mtune=native` is supported by the compiler. GCC
documentation says it should do nothing if unsupported by a particular
compiler but that doesn't seem to be the case.  Doesn't hurt to perform
yet another configure-time test.

Based on patch by @fischermario. Closes #154.