Allow multiline strings in configuration files
This uses a syntax similar to Python's multiline string, where three
single quotes (''') are used to mark the beginning and end of the
string.

The only limitation is that the opening quotes must be at the end of
a line, and the closing quotes must be by themselves in the last line.
For example:

	section name {
		something = '''
	this
		is
			a
				multiline
					string
	'''
	}